requirements.txt
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
xgboost==1.7.6
tensorflow==2.13.0
plotly==5.15.0
dash==2.11.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pytest==7.4.0
tqdm==4.65.0
joblib==1.3.1
PyYAML==6.0.1 